# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SSSD-APP API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://b-shop-api-5ntn.onrender.com/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/carts": {
            "get": {
                "tags": [
                    "Carts"
                ],
                "summary": "Get list of carts",
                "description": "Returns list of carts",
                "operationId": "bf4fcfa5c4963324af8f7c4709190694",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Cart"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Carts"
                ],
                "summary": "Create a new cart",
                "description": "Create a new cart",
                "operationId": "4166fd0418f4557cdaaf8dfcc29eb00d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Cart"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cart"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/carts/{id}": {
            "get": {
                "tags": [
                    "Carts"
                ],
                "summary": "Get specified cart",
                "description": "Returns specified cart",
                "operationId": "2a3c77b10b2ac68693ab31bd63102a98",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the cart",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cart"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Carts"
                ],
                "summary": "Update specified cart",
                "description": "Update specified cart",
                "operationId": "ab922d9c70155dfe7fc1f040ec36420d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the cart",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Cart"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cart"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Carts"
                ],
                "summary": "Delete specified cart",
                "description": "Delete specified cart",
                "operationId": "5f325f0730a847327428a8f5f80423a9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the cart",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get list of categories",
                "description": "Returns list of categories",
                "operationId": "ead5fb1cf2a7453acb2e37928a1b1eaf",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Category"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Create a new category",
                "description": "Create a new category",
                "operationId": "bb02f0320882db37f99574196577e446",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Category"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get specified category",
                "description": "Returns specified category",
                "operationId": "4cdf41756315109f1efc72787db5830a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Update specified category",
                "description": "Update specified category",
                "operationId": "19954b1f8aa4916b5f032dc28ea36a24",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Category"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Delete specified category",
                "description": "Delete specified category",
                "operationId": "f1b8af6e84bd52b9594d98bde748ebd6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Cart": {
                "title": "Cart",
                "description": "Cart model",
                "properties": {
                    "id": {
                        "description": "Unique identifier for the cart item",
                        "type": "integer",
                        "example": "1"
                    },
                    "user_id": {
                        "description": "User ID associated with the cart item",
                        "type": "integer",
                        "example": "1"
                    },
                    "product_id": {
                        "description": "Product ID associated with the cart item",
                        "type": "integer",
                        "example": "1"
                    },
                    "quantity": {
                        "description": "Quantity of the product in the cart",
                        "type": "integer",
                        "example": "2"
                    },
                    "created_at": {
                        "description": "Date and time when the cart item was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-12 10:00:00"
                    },
                    "updated_at": {
                        "description": "Date and time when the cart item was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-12 10:30:00"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "product": {
                        "$ref": "#/components/schemas/Product"
                    }
                },
                "type": "object"
            },
            "Category": {
                "title": "Category",
                "description": "Category model",
                "properties": {
                    "id": {
                        "description": "Unique identifier for the category",
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "description": "Name of the category",
                        "type": "string",
                        "example": "Electronics"
                    },
                    "description": {
                        "description": "Description of the category",
                        "type": "string",
                        "example": "Products related to electronics."
                    },
                    "created_at": {
                        "description": "Date and time when the category was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-12 09:00:00"
                    },
                    "updated_at": {
                        "description": "Date and time when the category was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-12 09:30:00"
                    },
                    "products": {
                        "description": "Products associated with the category",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Product"
                        }
                    }
                },
                "type": "object"
            },
...
            "Product": {
                "title": "Product",
                "description": "Product model",
                "properties": {
                    "id": {
                        "description": "Unique identifier for the product",
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "description": "Name of the product",
                        "type": "string",
                        "example": "Smartphone"
                    },
                    "description": {
                        "description": "Description of the product",
                        "type": "string",
                        "example": "A high-end smartphone with advanced features."
                    },
                    "price": {
                        "description": "Price of the product",
                        "type": "number",
                        "format": "float",
                        "example": "599.99"
                    },
                    "created_at": {
                        "description": "Date and time when the product was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-12 03:00:00"
                    },
                    "updated_at": {
                        "description": "Date and time when the product was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-12 03:30:00"
                    },
                    "category": {
                        "description": "Categories associated with the product",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    },
                    "reviews": {
                        "description": "Reviews associated with the product",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Review"
                        }
                    },
                    "carts": {
                        "description": "Carts associated with the product",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Cart"
                        }
                    }
                },
                "type": "object"
            },
            "Review": {
                "title": "Review",
                "description": "Review model",
                "properties": {
                    "id": {
                        "description": "Unique identifier for the review",
                        "type": "integer",
                        "example": "1"
                    },
                    "user_id": {
                        "description": "User ID associated with the review",
                        "type": "integer",
                        "example": "1"
                    },
                    "product_id": {
                        "description": "Product ID associated with the review",
                        "type": "integer",
                        "example": "1"
                    },
                    "rating": {
                        "description": "Rating given in the review",
                        "type": "integer",
                        "example": "4"
                    },
                    "review": {
                        "description": "Content of the review",
                        "type": "string",
                        "example": "This product is great!"
                    },
                    "created_at": {
                        "description": "Date and time when the review was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-12 02:00:00"
                    },
                    "updated_at": {
                        "description": "Date and time when the review was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-12 02:30:00"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "product": {
                        "$ref": "#/components/schemas/Product"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "description": "The unique identifier for the user",
                        "type": "integer",
                        "example": "1"
                    },
                    "full_name": {
                        "description": "The full name of the user",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "username": {
                        "description": "The username of the user",
                        "type": "string",
                        "example": "johndoe"
                    },
                    "email": {
                        "description": "The email address of the user",
                        "type": "string",
                        "format": "email",
                        "example": "johndoe@example.com"
                    },
                    "password": {
                        "description": "The password of the user (hashed)",
                        "type": "string",
                        "format": "password",
                        "example": "$2y$10$N0wbN10w.0.yBhEWwH2fveCAeAxFxuIQJ4O7QHBTQm7v9eN6LpfSi"
                    },
                    "phone_number": {
                        "description": "The phone number of the user",
                        "type": "string",
                        "example": "+1234567890"
                    },
                    "email_verified_at": {
                        "description": "The timestamp of email verification",
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-03-28T10:15:00Z"
                    },
                    "created_at": {
                        "description": "The timestamp of when the user was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-03-28T10:15:00Z"
                    },
                    "updated_at": {
                        "description": "The timestamp of when the user was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-03-28T10:15:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}